Progress on GNN-ACE.

+++ examples/ACE/utils/InteratomicBasisPotentialExtensions.jl
@@ -5,7 +5,7 @@ include("NNIAP.jl")
 Base.size(fd::ForceDescriptors) = (length(fd), )
 
 # Compute descriptors of a basis system and dataset
-function compute_local_descriptors(ds::DataSet, basis::BasisSystem; T = Float32)
+function compute_local_descriptors(ds::DataSet, basis::BasisSystem; T = Float64)
     e_des = Vector{LocalDescriptors}(undef, length(ds))
     #for (j, sys) in ProgressBar(zip(1:length(ds), get_system.(ds)))
     Threads.@threads for (j, sys) in ProgressBar(collect(enumerate(get_system.(ds))))
@@ -14,7 +14,7 @@ function compute_local_descriptors(ds::DataSet, basis::BasisSystem; T = Float32)
     return e_des
 end
 
-function compute_force_descriptors(ds::DataSet, basis::BasisSystem; T = Float32)
+function compute_force_descriptors(ds::DataSet, basis::BasisSystem; T = Float64)
     f_des = Vector{ForceDescriptors}(undef, length(ds))
     #for (j, sys) in ProgressBar(zip(1:length(ds), get_system.(ds)))
     Threads.@threads for (j, sys) in ProgressBar(collect(enumerate(get_system.(ds))))
